requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn
drf-spectacular==0.27.2
redis==5.0.8
celery==5.4.0
django-celery-beat==2.6.0
requests==2.32.3
django-unfold==0.39.0